chore: remove duplicated ComponentClassTypeHandler
